Remove babel load dependency, adds multer dependency. eslint correction for semi,space,..
