+++ package.json
@@ -4,41 +4,42 @@
   "description": "",
   "main": "./src/borderlineMiddleware.js",
   "scripts": {
-    "start": "node ./src/index.js",
-    "lint": "eslint src"
+      "start": "node ./src/index.js",
+      "lint": "eslint src"
   },
   "keywords": [
-    "borderline",
-    "etricks",
-    "zirg",
-    "ZirgMaster"
+      "borderline",
+      "middleware",
+      "etricks",
+      "zirg",
+      "ZirgMaster"
   ],
   "author": "Jean Grizet <jean.grizet16@imperial.ac.uk>",
   "license": "MIT",
   "readme": "README.md",
   "repository": {
-    "type": "git",
-    "url": "git+https://github.com/dsi-icl/borderline-middleware.git"
+      "type": "git",
+      "url": "git+https://github.com/dsi-icl/borderline-middleware.git"
   },
   "bugs": {
-    "url": "https://github.com/dsi-icl/borderline-middleware/issues"
+      "url": "https://github.com/dsi-icl/borderline-middleware/issues"
   },
   "homepage": "https://github.com/dsi-icl/borderline-middleware#readme",
   "devDependencies": {
-    "babel-core": "6.24.1",
-    "babel-eslint": "7.2.3",
-    "babel-loader": "7.0.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-stage-0": "6.24.1",
-    "eslint": "3.19.0"
+      "babel-core": "6.24.1",
+      "babel-eslint": "7.2.3",
+      "babel-preset-es2015": "6.24.1",
+      "babel-preset-stage-0": "6.24.1",
+      "eslint": "3.19.0"
   },
   "dependencies": {
-    "body-parser": "1.17.2",
-    "express": "4.15.3",
-    "fs-extra": "3.0.1",
-    "ip": "1.1.5",
-    "mongodb": "2.2.27",
-    "request": "2.81.0",
-    "uuid": "3.0.1"
+      "body-parser": "1.17.2",
+      "express": "4.15.3",
+      "fs-extra": "3.0.1",
+      "ip": "1.1.5",
+      "mongodb": "2.2.27",
+      "multer": "1.3.0",
+      "request": "2.81.0",
+      "uuid": "3.0.1"
   }
 }
